+++ jinahub/indexers/searcher/compound/NumpyFileSearcher/config.yml
@@ -3,6 +3,6 @@ with:
   {}
 metas:
   py_modules:  # THIS WILL NOT WORK SINCE IT RELIES ON CODE FILES FROM OTHER FOLDERS
-    - __init__.py
+    - npfile.py
   workspace: workspace/
   name: numpy_file_searcher
