+++ jinahub/indexers/searcher/compound/NumpyPostgresSearcher/config.yml
@@ -2,7 +2,7 @@ jtype: NumpyPostgresSearcher
 with:
   {}
 metas:
-  py_modules:  # THIS WILL NOT WORK SINCE IT RELIES ON CODE FILES FROM OTHER FOLDERS
-    - __init__.py
+  py_modules:
+    - nppsql.py
   workspace: workspace/
   name: numpy_postgres_searcher
