+++ jinahub/indexers/searcher/keyvalue/PostgreSQLSearcher/start.sh
